Backend/JapariService/requirements.txt
# python>=3.10,<3.12

requests
psycopg2-binary
Django>=4.2.0,<5.0.0
djangorestframework>=3.0.0,<4.0.0
daphne>=4.0.0,<5.0.0

# for testing
pytest-django
django-debug-toolbar
